0-requirements.txt
annotated-doc==0.0.4
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
causal-conv1d @ file:///hy-tmp/manba1/causal_conv1d-1.5.0.post7%2Bcu12torch2.4cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=11c449eb8822a6bcc4cb895f6151a2e9add260f53d6ace54e429342ec317568a
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.2
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
einops==0.8.2
executing==2.1.0
fair-esm==2.0.0
fastjsonschema==2.20.0
filelock==3.25.2
fqdn==1.5.1
fsspec==2026.2.0
h11==0.14.0
hf-xet==1.4.2
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface_hub==0.36.2
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.5.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-language-pack-zh-CN==4.2.post3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
mamba-ssm @ file:///hy-tmp/manba1/mamba_ssm-2.2.3.post2%2Bcu12torch2.4cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=d0f726b4c751eef8bfb1344c5ab5d1262029328f530e3a9766fc3afeef68b3d0
markdown-it-py==4.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
notebook_shim==0.2.4
numpy==2.4.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-contrib-python==4.10.0.84
overrides==7.7.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py3nvml==0.2.7
pycparser==2.22
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
PyYAML==6.0.3
pyzmq==26.2.0
referencing==0.35.1
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.20.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.3.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.5
sniffio==1.3.1
soupsieve==2.6
ssh-import-id==5.11
stack-data==0.6.3
supervisor==4.2.5
sympy==1.12
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tornado==6.4.1
tqdm==4.67.3
traitlets==5.14.3
transformers==4.40.0
triton==3.0.0
typer==0.24.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.15.0
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.6.3
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
xmltodict==0.14.1
zipp==1.0.0
